🚮 Get rid of the second beta role

+++ src/modules/betaAccess/events/guildMemberUpdate.ts
@@ -1,7 +1,7 @@
 import * as Discord from "discord.js";
 import { MongoClient } from "../../../database/client";
 
-var { betaTester, patron, beta, booster } = require("../../../roles.json"),
+var { patron, beta, booster, donator } = require("../../../roles.json"),
   coll = MongoClient.db("PreMiD").collection("betaAccess");
 
 module.exports = async (
@@ -9,11 +9,8 @@ module.exports = async (
   newMember: Discord.GuildMember
 ) => {
   //* If user is patron and does not have either betaTester or beta role, give it to them.
-  if (
-    newMember.roles.has(patron) &&
-    (!newMember.roles.has(betaTester) || !newMember.roles.has(beta))
-  ) {
-    newMember.roles.add([betaTester, beta]);
+  if (newMember.roles.has(patron) && !newMember.roles.has(beta)) {
+    newMember.roles.add(beta);
 
     if (!(await coll.findOne({ userId: newMember.id })))
       await coll.insertOne({ userId: newMember.id });
@@ -23,7 +20,7 @@ module.exports = async (
 
   //* If user boosts and doesn't have beta role, give it to them.
   if (newMember.roles.has(booster) && !newMember.roles.has(beta)) {
-    newMember.roles.add([beta, betaTester]);
+    newMember.roles.add(beta);
 
     coll.insertOne({ userId: newMember.id });
 
@@ -31,30 +28,23 @@ module.exports = async (
   }
 
   //* Remove beta access when boost expires.
-  if (oldMember.roles.has(booster) && !newMember.roles.has(booster)) {
+  if (
+    oldMember.roles.has(booster) &&
+    !newMember.roles.has(booster) &&
+    !newMember.roles.has(patron) &&
+    !newMember.roles.has(donator)
+  ) {
     newMember.roles.remove(beta);
-    if (!oldMember.roles.has(patron)) newMember.roles.remove(betaTester);
+    if (!oldMember.roles.has(patron)) newMember.roles.remove(beta);
 
     coll.findOneAndDelete({ userId: newMember.id });
 
     return;
   }
-  
+
   //* Remove beta access when the beta role is removed.
-  if (oldMember.roles.has(beta) && !newMember.roles.has(beta) && newMember.roles.has(betaTester)) {
-    newMember.roles.remove(betaTester);
-    
+  if (oldMember.roles.has(beta) && !newMember.roles.has(beta)) {
     coll.findOneAndDelete({ userId: newMember.id });
-
-    return;
-  }
-  
-  //* Remove beta access when the beta tester is removed.
-  if (oldMember.roles.has(betaTester) && !newMember.roles.has(betaTester) && newMember.roles.has(beta)) {
-    newMember.roles.remove(beta);
-    
-    coll.findOneAndDelete({ userId: newMember.id });
-
     return;
   }
 };
